% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_humanization.R
\name{data_humanization}
\alias{data_humanization}
\title{Humanize phosphoproteomic data}
\usage{
data_humanization(phospho_df = phospho, output_folder = "data/")
}
\arguments{
\item{phospho_df}{\verb{<dataframe>} Dataframe to humanize}

\item{output_folder}{\verb{<character>} Where the output files should be stored}
}
\description{
This function humanizes mouse phosphoproteomic data according to the
PhosphoSitePlus orthology database.
}
