% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_setup.R
\name{dependencies_setup}
\alias{dependencies_setup}
\title{Dependencies setup}
\usage{
dependencies_setup(networkin = TRUE, ivkea = TRUE)
}
\arguments{
\item{networkin}{\verb{<bolean>} Whether to install NetworKIN or not}

\item{ivkea}{\verb{<bolean>} Whether to install IV-KEA or not}
}
\description{
This script installs IVKEA and NetworKIN in the current directory
}
