% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_data.R
\name{dependency_data}
\alias{dependency_data}
\title{Data dependency analysis}
\usage{
dependency_data(dependency_kinases_output, functional_threshold = 0.25)
}
\arguments{
\item{dependency_kinases_output}{\verb{<dataframe>} Output table of the dependency_kinases function}

\item{functional_threshold}{\verb{<integer>} Minimum threshold to filter functional
phosphosites as defined by Ochoa et al., 2019 (0 is a not functional site, 1 a highly functional site)}
}
\description{
Fetch raw dependency data for a given set of kinase-cell lineage combination. Useful for visualization
and to identify dependent proteins targeted by tested kinases.
}
