% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_graph.R
\name{dependency_graph}
\alias{dependency_graph}
\title{Dependency analysis visualization}
\usage{
dependency_graph(dependency_data_output)
}
\arguments{
\item{dependency_data_output}{\verb{<dataframe>} Output of the dependency_data function}
}
\description{
Draw a volcano plot from any test results
}
