% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_kinases.R
\name{dependency_kinases}
\alias{dependency_kinases}
\title{Kinases dependency analysis}
\usage{
dependency_kinases(
  kinases,
  dependency_threshold = -0.2,
  functional_threshold = 0.25
)
}
\arguments{
\item{kinases}{\verb{<vector>} Vector of kinases to analyze}

\item{dependency_threshold}{\verb{<integer>} Minimum dependency score to consider
(0 is a non-essentiel gene, 1 an essential gene)}

\item{functional_threshold}{\verb{<integer>} Minimum threshold to filter functional
phosphosites as defined by Ochoa et al., 2019 (0 is a not functional site, 1 a highly functional site)}
}
\description{
Computes the dependency score of the given list of kinases using DepMap RNAi data.
}
