% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniprot_ids.R
\name{.get_uniprot_id}
\alias{.get_uniprot_id}
\title{get_uniprot_id
This functions outputs a dataframe containing
the queried UniprotKB names and its corresponding
UniprotKB accession IDs.}
\usage{
.get_uniprot_id(query_names)
}
\arguments{
\item{query_names}{\verb{<vector>}}
}
\description{
get_uniprot_id
This functions outputs a dataframe containing
the queried UniprotKB names and its corresponding
UniprotKB accession IDs.
}
