% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predictions.R
\name{filter_predictions}
\alias{filter_predictions}
\title{Filter NetworKIN predictions}
\usage{
filter_predictions(
  predictions_file = "data/outputs/networKIN_output.tsv",
  threshold = 0.8,
  experiment = "test"
)
}
\arguments{
\item{predictions_file}{\verb{<character>} Location of NetworKIN output file}

\item{threshold}{\verb{<numeric>} NetworKIN prediction score threshold used}

\item{experiment}{\verb{<character>} Name of the experiment to tag output files}
}
\description{
NetworKIN predictions are filtered to keep the top 20\% prediction scores.
}
