% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniprot_ids.R
\name{get_uniprot_ids}
\alias{get_uniprot_ids}
\title{get_uniprot_ids
This function outputs a dataframe of UniprotKB accession IDs
and UniprotKB entry names. The dataframe is sliced to submit a
precise number of IDs per request (to avoid too long URLs).
UniprotKB entry names are submitted as a UniprotKB search to access
the corresponding UniprotKB IDs.}
\usage{
get_uniprot_ids(uniprot_names, query_length = 750)
}
\arguments{
\item{uniprot_names}{\verb{<vector>} UniprotKB names}

\item{query_length}{\verb{<integer>} The number of IDs to submit in once.
Change this parameter only if a HTTP404 error is outputted.}
}
\description{
get_uniprot_ids
This function outputs a dataframe of UniprotKB accession IDs
and UniprotKB entry names. The dataframe is sliced to submit a
precise number of IDs per request (to avoid too long URLs).
UniprotKB entry names are submitted as a UniprotKB search to access
the corresponding UniprotKB IDs.
}
