% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitrodb.R
\docType{data}
\name{invitrodb}
\alias{invitrodb}
\title{In vitro database for IV-KEA}
\format{
An object of class \code{"data.frame"}
}
\usage{
data("invitrodb")
}
\description{
This database is used to perform IV-KEA. It indicates all phosphosites
phosphorylated by a set of 350 kinases according to an in vitro experiment.
}
\examples{
data("invitrodb") 
head(invitrodb)
}
\references{
This database was generated by Sugiyama et al., 2019
}
\keyword{dataset}
