% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivkea_setup.R
\name{ivkea_setup}
\alias{ivkea_setup}
\title{IV-KEA setup
This function installs the in vitro database from Sugiyama et al. 2019.
It downloads all UniprotKB accession IDs that are required for IV-KEA.
This function is not part of the main pipeline as the projet already includes
a ready-to-use in vitro database.}
\usage{
ivkea_setup(
  invitrodb_file = "data/imported/db_phosphosite_kinase_invitro.xlsx"
)
}
\arguments{
\item{invitrodb_file}{\verb{<character>} Location of the in vitro database
from Sugiyama 2019.}
}
\description{
IV-KEA setup
This function installs the in vitro database from Sugiyama et al. 2019.
It downloads all UniprotKB accession IDs that are required for IV-KEA.
This function is not part of the main pipeline as the projet already includes
a ready-to-use in vitro database.
}
