% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_volcano_plot.R
\name{make_volcano_plot}
\alias{make_volcano_plot}
\title{Draw a volcano plot from any test results}
\usage{
make_volcano_plot(
  kinase_enrichment_file = "data/analyses/networkin_kinase_enrichment.csv",
  odds_ratio = odds_ratio,
  FDR_cutoff = 0.05,
  FDR_data = FDR,
  labels = top_predicted_kinase,
  graph_title = "Volcano plot",
  x_axis_title = "X axis",
  save_path
)
}
\arguments{
\item{kinase_enrichment_file}{\verb{<character>} Location of the file with Fisher exact test results}

\item{odds_ratio}{\verb{<column_name>} Unquoted name of the \code{kinase_enrichment_file} column
that contains the odds ratios.}

\item{FDR_cutoff}{\verb{<numeric>} Cutoff to display.}

\item{FDR_data}{\verb{<column_name>} Unquoted name of \code{kinase_enrichment_file} column
that contains the FDR values.}

\item{labels}{\verb{<column_name>} Unquoted name of \code{kinase_enrichment_file} column
that contains the labels.}

\item{graph_title}{\verb{<character>}}

\item{x_axis_title}{\verb{<character>}}

\item{save_path}{\verb{<character>} File path to save the volcano plot in.}
}
\description{
Draw a volcano plot from any test results
}
