% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkin_input.R
\name{networkin_input}
\alias{networkin_input}
\title{Generate NetworKIN input
This function imports, clean, and prepare phosphoproteomic
data for NetworKIN.}
\usage{
networkin_input(
  phospho_file = "data/imported/phospho.xlsx",
  species = "hsa",
  phosphosites_column = "ProteinID-Phospho:Site",
  log2_column = "Log2(Relapse/diagnostic)",
  fdr_column = "Adj. p-value",
  experiment = "test"
)
}
\arguments{
\item{phospho_file}{\verb{<.xlsx file>} Raw phosphoproteomic data}

\item{species}{\verb{<character>} Species code from which samples were obtained (\code{hsa} or \code{mmu}).}

\item{phosphosites_column}{\verb{<column_name>} Column with phosphosite names}

\item{log2_column}{\verb{<column_name>} Column with log2(experiment/control) values}

\item{fdr_column}{\verb{<column_name>} Column with adjusted pvalues}

\item{experiment}{\verb{<character>} Name of the experiment to tag output files}
}
\description{
Generate NetworKIN input
This function imports, clean, and prepare phosphoproteomic
data for NetworKIN.
}
