% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkin_qc.R
\name{networkin_qc}
\alias{networkin_qc}
\title{Quality control of NetworKIN predictions}
\usage{
networkin_qc(
  predictions_file = "networkin_output.csv",
  threshold = 0.5,
  output_folder = "myexperiment/"
)
}
\arguments{
\item{predictions_file}{\verb{<character>} Location of NetworKIN output file}

\item{threshold}{\verb{<numeric>} NetworKIN prediction score threshold used}

\item{output_folder}{\verb{<character>} Where the output files should be stored}
}
\description{
Draw a density plot of all predictions made by NetworKIN before filtering
}
