% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkin_setup.R
\name{networkin_setup}
\alias{networkin_setup}
\title{NetworKIN setup
NetworKIN is properly downloaded and installed with this function.}
\usage{
networkin_setup()
}
\description{
NetworKIN setup
NetworKIN is properly downloaded and installed with this function.
}
