% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ivkea.R
\name{perform_ivkea}
\alias{perform_ivkea}
\title{Perform IV-KEA
This function performs IV-KEA on a given phosphoproteomic file
according to the in vitro database produced by Sugiyama et al. 2019.}
\usage{
perform_ivkea(
  clean_phospho_file = "phospho_clean.csv",
  output_folder = "myexperiment/"
)
}
\arguments{
\item{clean_phospho_file}{\verb{<character>} Location of the clean
phosphoproteomic file}

\item{output_folder}{\verb{<character>} Where the output files should be stored}
}
\description{
Perform IV-KEA
This function performs IV-KEA on a given phosphoproteomic file
according to the in vitro database produced by Sugiyama et al. 2019.
}
