% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_input.R
\name{phospho_input}
\alias{phospho_input}
\title{Generate a clean phosphoproteomic file for NetworKIN and IV-KEA}
\usage{
phospho_input(
  phospho_file = "phospho_human.xlsx",
  species = "hsa",
  phosphosites_column = "PhosphoSite",
  log2_column = "Log2",
  fdr_column = "Adj. Pvalue",
  output_folder = "myexperiment/"
)
}
\arguments{
\item{phospho_file}{\verb{<.xlsx or .csv file>} Raw phosphoproteomic data}

\item{species}{\verb{<character>} Species code from which samples were obtained (\code{hsa} or \code{mmu}).}

\item{phosphosites_column}{\verb{<column_name>} Column with phosphosite names}

\item{log2_column}{\verb{<column_name>} Column with log2(experiment/control) values}

\item{fdr_column}{\verb{<column_name>} Column with adjusted pvalues}

\item{output_folder}{\verb{<character>} Where the output files should be stored (with a \code{/} at the end)}
}
\description{
This function imports, clean, and prepare phosphoproteomic
data for NetworKIN and IV-KEA.
}
