% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_comparison.R
\name{predictions_comparison}
\alias{predictions_comparison}
\title{Compare predictions}
\usage{
predictions_comparison(
  ivkea_enrichment_file,
  networkin_enrichment_file,
  FDR_cutoff = 0.05,
  graph_title = "Kinases prediction enrichment Networkin vs IV-KEA",
  output_folder = "myexperiment/",
  file_name
)
}
\arguments{
\item{ivkea_enrichment_file}{\verb{<character>} Location of the IV-KEA
file with Fisher exact test results}

\item{networkin_enrichment_file}{\verb{<character>} Location of the IV-KEA
file with Fisher exact test results}

\item{FDR_cutoff}{\verb{<numeric>} Cutoff to display.}

\item{graph_title}{\verb{<character>}}

\item{output_folder}{\verb{<character>} Where the output files should be stored}

\item{file_name}{\verb{<character>} File name including extension (.pdf recommended)}
}
\description{
Print NetworKIN and IV-KEA predictions on the same volcano plot
to compare predictions.
}
