% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_networkin.R
\name{run_networkin}
\alias{run_networkin}
\title{Run NetworKIN
This function prepares and executes NetworKIN to generate predictions
on a given set of phosphorylated peptides.
NetworKIN requires Python and the BLASTALL version of BLAST.}
\usage{
run_networkin(
  input_file = "data/outputs/networKIN_input_test.res",
  blastall_location,
  experiment = "test"
)
}
\arguments{
\item{input_file}{\verb{<character>} Location of the input file}

\item{blastall_location}{\verb{<character>} Location of BLASTALL on the computer}

\item{experiment}{\verb{<character>} Name of the experiment to tag output}
}
\description{
Run NetworKIN
This function prepares and executes NetworKIN to generate predictions
on a given set of phosphorylated peptides.
NetworKIN requires Python and the BLASTALL version of BLAST.
}
