% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_top_predictions.R
\name{select_top_predictions}
\alias{select_top_predictions}
\title{Extract top NetworKIN predictions
The strongest NetworKIN predictions for each phosphosite are selected.}
\usage{
select_top_predictions(
  predictions_file = "data/analyses/filtered_networkin_predictions.csv",
  phospho_cleaned_file = "data/outputs/phospho_clean.csv",
  experiment = "test"
)
}
\arguments{
\item{predictions_file}{\verb{<character>} Location of the filtered NetworKIN output file}

\item{phospho_cleaned_file}{\verb{<character>} Location of the cleaned phosphoproteomic file
generated by \code{networkin_input}.}

\item{experiment}{\verb{<character>} Name of the experiment to tag output files}
}
\description{
Extract top NetworKIN predictions
The strongest NetworKIN predictions for each phosphosite are selected.
}
