% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivkea_setup.R
\name{ivkea_setup}
\alias{ivkea_setup}
\title{IV-KEA setup}
\usage{
ivkea_setup(ivkea_folder = "data/imported/")
}
\arguments{
\item{ivkea_folder}{\verb{<character>} Where to download IV-KEA database}
}
\description{
IV-KEA database is downloaded. Run this function only for custom IV-KEA setup.
}
