% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_fisher_exact_test.R
\name{perform_Fisher_exact_test}
\alias{perform_Fisher_exact_test}
\title{Perform Fisher exact test}
\usage{
perform_Fisher_exact_test(
  top_predictions_file = "data/analyses/top_predictions.csv",
  experiment = "test",
  predicted_kinase
)
}
\arguments{
\item{top_predictions_file}{\verb{<character>} Location of the file with top predictions per phosphosite}

\item{experiment}{\verb{<character>} Name of the experiment to tag output files}
}
\description{
This function computes Fisher exact test and the corresponding adjusted p-values (FDR)
for upregulated vs downregulated phosphosites, upregulated vs all phosphosites,
downregulated vs all phosphosites.
}
