% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ivkea.R
\name{perform_ivkea}
\alias{perform_ivkea}
\title{Perform IV-KEA
This function performs IV-KEA on a given phosphoproteomic file
according to the in vitro database produced by Sugiyama et al. 2019.}
\usage{
perform_ivkea(
  clean_phospho_file = "data/outputs/phospho_clean.csv",
  invitrodb_file = "data/imported/invitrodb.csv",
  experiment = "test"
)
}
\arguments{
\item{clean_phospho_file}{\verb{<character>} Location of the clean
phosphoproteomic file}

\item{invitrodb_file}{\verb{<character>} Location of the in vitro database.
This must be the ready-to-use version of the database produced by \code{ivkea_setup}}

\item{experiment}{\verb{<character>} Name of the experiment to tag output}

\item{uniprot_id_phospho}{\verb{<column_name>} Column with UniprotKB IDs in
the file containing cleaned phosphoproteomic results.}

\item{uniprot_id_invitrodb}{Column with UniprotKB IDs in the in vitro
database from Sugiyama 2019.}
}
\description{
Perform IV-KEA
This function performs IV-KEA on a given phosphoproteomic file
according to the in vitro database produced by Sugiyama et al. 2019.
}
