% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_comparison.R
\name{predictions_comparison}
\alias{predictions_comparison}
\title{Compare predictions}
\usage{
predictions_comparison(
  ivkea_enrichment_file = "data/analyses/ivkea_kinase_enrichment.csv",
  networkin_enrichment_file = "data/analyses/networkin_kinase_enrichment.csv",
  FDR_cutoff = 0.05,
  graph_title = "Kinases prediction enrichment Networkin vs IV-KEA",
  save_path = "figures/predictions_comparison.pdf"
)
}
\arguments{
\item{ivkea_enrichment_file}{\verb{<character>} Location of the IV-KEA
file with Fisher exact test results}

\item{networkin_enrichment_file}{\verb{<character>} Location of the IV-KEA
file with Fisher exact test results}

\item{FDR_cutoff}{\verb{<numeric>} Cutoff to display.}

\item{graph_title}{\verb{<character>}}

\item{save_path}{\verb{<character>} File path to save the volcano plot in.}
}
\description{
Print NetworKIN and IV-KEA predictions on the same volcano plot
to compare predictions.
}
