% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_networkin.R
\name{run_networkin}
\alias{run_networkin}
\title{Run NetworKIN}
\usage{
run_networkin(
  input_file = "networKIN_input.res",
  blastall_location,
  output_folder = "myexperiment/"
)
}
\arguments{
\item{input_file}{\verb{<character>} Location of the input file}

\item{blastall_location}{\verb{<character>} Location of BLASTALL on the computer}

\item{output_folder}{\verb{<character>} Where the output files should be stored}
}
\description{
This function prepares and executes NetworKIN to generate predictions
on a given set of phosphorylated peptides.
NetworKIN requires Python and the BLASTALL version of BLAST.
}
