% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_networkin.R
\name{run_networkin}
\alias{run_networkin}
\title{Run NetworKIN}
\usage{
run_networkin(
  input_file = "phospho_clean.csv",
  output_folder = "myexperiment/"
)
}
\arguments{
\item{input_file}{\verb{<character>} Location of \code{phospho_clean} file}

\item{output_folder}{\verb{<character>} Where the output files should be stored}
}
\description{
This function uses pre-computed NetworKIN predictions to generate predictions
on a given set of phosphorylated peptides.
}
