% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_top_predictions.R
\name{select_top_predictions}
\alias{select_top_predictions}
\title{Extract top NetworKIN predictions}
\usage{
select_top_predictions(
  predictions_file = "networKIN_output.tsv",
  phospho_cleaned_file = "phospho_clean.csv",
  output_folder = "data/"
)
}
\arguments{
\item{predictions_file}{\verb{<character>} Filename of the  NetworKIN output file}

\item{phospho_cleaned_file}{\verb{<character>} Filename of the cleaned phosphoproteomic file
generated by \code{networkin_input}.}

\item{output_folder}{\verb{<character>} Where the output files should be stored}
}
\description{
The strongest NetworKIN predictions for each phosphosite are selected.
}
